installed navigation packages and refactored the code from App.js to Home.js
